[refactor] #96 dto 패키지 분리
