+++ src/main/java/kyonggi/bookslyserver/domain/shop/dto/request/employee/EmployeeWorkScheduleDto.java
@@ -1,4 +1,4 @@
-package kyonggi.bookslyserver.domain.shop.dto.request;
+package kyonggi.bookslyserver.domain.shop.dto.request.employee;
 
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
