+++ src/main/java/kyonggi/bookslyserver/domain/shop/dto/request/menu/MenuCategoryCreateDto.java
@@ -1,4 +1,4 @@
-package kyonggi.bookslyserver.domain.shop.dto.request;
+package kyonggi.bookslyserver.domain.shop.dto.request.menu;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
