+++ src/main/java/kyonggi/bookslyserver/domain/shop/dto/response/shop/ShopCreateResponseDto.java
@@ -1,6 +1,5 @@
 package kyonggi.bookslyserver.domain.shop.dto.response.shop;
 
-import kyonggi.bookslyserver.domain.shop.dto.BusinessScheduleDto;
 import kyonggi.bookslyserver.domain.shop.entity.Shop.Shop;
 import kyonggi.bookslyserver.domain.shop.entity.Shop.ShopImage;
 import lombok.Data;
