+++ src/main/java/kyonggi/bookslyserver/domain/shop/entity/Employee/Employee.java
@@ -3,7 +3,7 @@ package kyonggi.bookslyserver.domain.shop.entity.Employee;
 import jakarta.persistence.*;
 import kyonggi.bookslyserver.domain.event.entity.timeEvent.EmployeeTimeEvent;
 import kyonggi.bookslyserver.domain.review.entity.Review;
-import kyonggi.bookslyserver.domain.shop.dto.request.EmployeeCreateRequestDto;
+import kyonggi.bookslyserver.domain.shop.dto.request.employee.EmployeeCreateRequestDto;
 import kyonggi.bookslyserver.domain.shop.entity.Menu.Menu;
 import kyonggi.bookslyserver.domain.shop.entity.Shop.Shop;
 import kyonggi.bookslyserver.global.common.BaseTimeEntity;
